requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alabaster==1.0.0
-e git+https://github.com/Aditya-gam/algolib-adg.git@2987f6f7513ec821eb12070a5a0434f3526ce1c0#egg=algolib_adg
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.2
dataclasses-json==0.6.7
distlib==0.3.9
docutils==0.21.2
filelock==3.18.0
frozenlist==1.7.0
furo==2024.8.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hypothesis==6.135.24
identify==2.6.12
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.45.0
nodeenv==1.9.1
numpy==2.3.1
ollama==0.5.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
pytest-benchmark==5.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
rpds-py==0.26.0
ruff==0.12.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinx-autobuild==2024.10.3
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
starlette==0.47.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
